[Refactor] Tests for DeveloperPortal::Admin::Account::PasswordsController in the same file
